Move query params into MapPage
