+++ src/app/pages/MapPage/map-search/MapSearchMarker.tsx
@@ -3,7 +3,7 @@ import { useMapInstance } from '@datapunt/react-maps'
 import { LeafletMouseEvent } from 'leaflet'
 import React, { useContext, useEffect } from 'react'
 import fetchNearestDetail from '../../../../map/services/nearest-detail/nearest-detail'
-import { detailUrlParam, locationParam } from '../../../query-params'
+import { detailUrlParam, locationParam } from '../query-params'
 import joinUrl from '../../../utils/joinUrl'
 import useParam from '../../../utils/useParam'
 import MapContext, { Overlay } from '../MapContext'
